fix(migrations): type the GVWR-class and duty-class columns integer, as their entities map them
pos-catalog's Postgres slices failed on b04c2130 with Hibernate's schema
validation: ext_skill.max_gvwr_class is smallint (int2) while
ExtSkillReplica maps an int, and validate refuses the width difference.
The same pairing was in every column this branch added for a class 1-8:
ext_skill and service_skill_requirement (pos-catalog), skill (pos-people),
ext_vehicle.gvwr_class and ext_catalog_service_skill (pos-shop-manager),
vehicle_records.gvwr_class (pos-vehicle-inventory), and max_duty_class on
bays and both ext_bay replicas. All nine columns are integer now, the way
V9__person_credential_replica already types the credential's range and the
way every entity declares the field. The migrations are this branch's own,
so no deployed schema carries the smallint.

The H2 unit slices never validate the schema, which is why this only showed
in CI; the CHECK constraints (1..8) are unchanged.

+++ pos-location/src/main/resources/db/migration/V3__bays_specialty_codes_and_duty_class.sql
@@ -22,7 +22,7 @@ UPDATE public.bays SET service_capability_codes = NULL;
 
 ALTER TABLE public.bays DROP COLUMN skill_requirement_ids;
 
-ALTER TABLE public.bays ADD COLUMN max_duty_class smallint;
+ALTER TABLE public.bays ADD COLUMN max_duty_class integer;
 ALTER TABLE public.bays
     ADD CONSTRAINT bays_max_duty_class_check
     CHECK (max_duty_class IS NULL OR max_duty_class BETWEEN 1 AND 8);
